+++ src/js/core.js
@@ -2801,7 +2801,8 @@ $(document).ready(function ($) {
                         anime: $('#anime').prop('checked'),
                         scene: $('#scene').prop('checked'),
                         defaultStatusAfter: $('#statusSelectAfter').val(),
-                        skip_downloaded: $('#skip_downloaded').prop('checked')
+                        skip_downloaded: $('#skip_downloaded').prop('checked'),
+                        add_show_year: $('#add_show_year').prop('checked')
                     });
 
                     $(this).attr('disabled', true);
@@ -2809,7 +2810,7 @@ $(document).ready(function ($) {
                     SICKRAGE.notify('info', gt('Saved Defaults'), gt('Your "add show" defaults have been set to your current selections.'));
                 });
 
-                $('#statusSelect, #qualityPreset, #flatten_folders, #anyQualities, #bestQualities, #subtitles, #scene, #anime, #statusSelectAfter, #skip_downloaded').change(function () {
+                $('#statusSelect, #qualityPreset, #flatten_folders, #anyQualities, #bestQualities, #subtitles, #scene, #anime, #statusSelectAfter, #skip_downloaded', '#add_show_year').change(function () {
                     $('#saveDefaultsButton').attr('disabled', false);
                 });
             },